// src/features/AdminDashboard/users/users.types.ts
export interface IUser {
  id: string;
  name: string | null;
  email: string;
  username: string;
  branchId: string | null;
  password: string | null;
  instituteName: string | null;
  directorName: string | null;
  mobileNumber: string | null;
  gender: string | null;
  nationality: string | null;
  instituteAge: string | null;
  religion: string | null;
  fullAddress: string | null;
  village: string | null;
  postOffice: string | null;
  thanaUpazila: string | null;
  district: string | null;
  fatherName: string | null;
  motherName: string | null;
  courseName: string | null;
  duration: string | null;
  startYear: string | null;
  startMonth: string | null;
  endYear: string | null;
  endMonth: string | null;
  educationQualification: string | null;
  directorPhoto: string | null;
  institutePhoto: string | null;
  nationalIdPhoto: string | null;
  signaturePhoto: string | null;
  emailVerified: boolean;
  needPasswordChange: boolean;
  isDeleted: boolean;
  deletedAt: string | null;
  role: "ADMIN" | "USER";
  status: "PENDING" | "ACTIVE" | "BLOCKED" | "DELETED";
  createdAt: string;
  updatedAt: string;
}